Hide isEntitledTo from public API
