+++ src/helpers/entitlement-checking-helper.ts
@@ -1,7 +1,8 @@
-import { Purchases } from "../main";
+import { EntitlementResponse } from "../networking/responses/entitlements-response";
+import { Backend } from "../networking/backend";
 
 export function waitForEntitlement(
-  purchases: Purchases,
+  backend: Backend,
   appUserId: string,
   entitlementIdentifier: string,
   maxAttempts: number = 10,
@@ -9,8 +10,7 @@ export function waitForEntitlement(
   const waitMSBetweenAttempts = 1000;
   return new Promise<boolean>((resolve, reject) => {
     const checkForEntitlement = (checkCount = 1) =>
-      purchases
-        .isEntitledTo(appUserId, entitlementIdentifier)
+      isEntitledTo(backend, appUserId, entitlementIdentifier)
         .then((hasEntitlement) => {
           if (checkCount > maxAttempts) {
             return resolve(false);
@@ -29,4 +29,17 @@ export function waitForEntitlement(
 
     checkForEntitlement();
   });
+}
+
+async function isEntitledTo(
+  backend: Backend,
+  appUserId: string,
+  entitlementIdentifier: string,
+): Promise<boolean> {
+  const entitlementsResponse = await backend.getEntitlements(appUserId);
+
+  const entitlements = entitlementsResponse.entitlements.map(
+    (ent: EntitlementResponse) => ent.lookup_key,
+  );
+  return entitlements.includes(entitlementIdentifier);
 }
